Fix db stuff

+++ tariff-management-service/app/src/main/java/com/microservices/tariffmanagement/entity/TariffEntity.java
@@ -4,11 +4,8 @@ import com.microservices.tariffmanagement.dto.TariffDto;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
+import javax.persistence.*;
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "tariffs")
@@ -21,14 +18,14 @@ public class TariffEntity {
 
     private String name;
 
-    private double price;
+    private BigDecimal price;
 
     //in millis
     private long washingTime;
 
     public TariffEntity(TariffDto tariffDto) {
         this.name = tariffDto.getName();
-        this.price = tariffDto.getPrice();
+        this.price = BigDecimal.valueOf(tariffDto.getPrice());
         this.washingTime = tariffDto.getWashingTime();
     }
 }
